code - extract in widget files of what could be done Fixes #53
